refactor: decouple stored session deletion from ChatRecordingService
Session deletion only needs `config.storage.getProjectTempDir()`, but the
CLI delete utility was constructing a `ChatRecordingService` with `Config`
to call `deleteSession`. ChatRecordingService's constructor expects an
`AgentLoopContext`, and the call only worked because `Config` implements
`AgentLoopContext` as a deprecated backwards-compat surface.

This extracts deletion into `deleteStoredSession(config, sessionIdOrBasename)`
in `utils/sessionOperations.ts` along with three helpers
(`deriveSessionShortId`, `getMatchingSessionFiles`, `deleteSessionFileAndArtifacts`).
`ChatRecordingService.deleteSession()` becomes a thin wrapper that delegates
to the free function, preserving the existing instance-method consumer in
`useSessionBrowser` and the test suite.

The CLI delete flow in `packages/cli/src/utils/sessions.ts` now calls the
free function directly, removing reliance on Config-as-AgentLoopContext.

Refs: https://github.com/google-gemini/gemini-cli/issues/22920
